feat: implement login form and signup page
- Added LoginForm widget for user authentication.
- Created SignupPage to handle user registration.
- Updated LoginSignupPage to include the new SignupPage and LoginForm.
