+++ lib/pages/login_signup_page.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
+import 'package:balance_ai_agent/widgets/login_form.dart';
 import 'chat_room_page.dart';
+import 'signup_page.dart';
 
 class LoginSignupPage extends StatelessWidget {
   const LoginSignupPage({super.key});
@@ -34,7 +36,6 @@ class LoginSignupPage extends StatelessWidget {
             const SizedBox(height: 16),
             SizedBox(
               width: 300,
-              height: 56,
               child: ElevatedButton(
                 onPressed: () {
                   // 匿名で使用する処理
@@ -54,8 +55,9 @@ class LoginSignupPage extends StatelessWidget {
             Row(
               mainAxisAlignment: MainAxisAlignment.center,
               children: [
-                Text("Does not have account?", style: TextStyle(fontSize: 16)),
-                SizedBox(width: 16),
+                const Text("Does not have account?",
+                    style: TextStyle(fontSize: 16)),
+                const SizedBox(width: 16),
                 InkWell(
                   // GestureDetector で Text をラップ
                   onTap: () {
@@ -79,111 +81,5 @@ class LoginSignupPage extends StatelessWidget {
         ),
       ),
     );
-  }
-}
-
-class LoginForm extends StatefulWidget {
-  const LoginForm({super.key});
-
-  @override
-  State<LoginForm> createState() => _LoginFormState();
-}
-
-class _LoginFormState extends State<LoginForm> {
-  final _formKey = GlobalKey<FormState>();
-  bool _obscurePassword = true;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(16.0),
-      child: Form(
-        key: _formKey,
-        child: SizedBox(
-          width: 300, // 幅を指定
-          child: Column(
-            children: [
-              TextFormField(
-                autofocus: true,
-                cursorColor: const Color.fromARGB(255, 93, 190, 163),
-                decoration: const InputDecoration(
-                  labelText: 'Email',
-                  border: OutlineInputBorder(),
-                ),
-                validator: (value) {
-                  if (value == null || value.isEmpty) {
-                    return 'Please enter your email';
-                  }
-                  if (!value.contains('@')) {
-                    return 'Please enter a valid email';
-                  }
-                  return null;
-                },
-              ),
-              const SizedBox(height: 16),
-              TextFormField(
-                obscureText: _obscurePassword,
-                cursorColor: const Color.fromARGB(255, 93, 190, 163),
-                decoration: InputDecoration(
-                  labelText: 'Password',
-                  border: const OutlineInputBorder(),
-                  suffixIcon: IconButton(
-                    icon: Icon(
-                      _obscurePassword
-                          ? Icons.visibility_off
-                          : Icons.visibility,
-                    ),
-                    onPressed: () {
-                      setState(() {
-                        _obscurePassword = !_obscurePassword;
-                      });
-                    },
-                  ),
-                ),
-                validator: (value) {
-                  if (value == null || value.isEmpty) {
-                    return 'Please enter your password';
-                  }
-                  return null;
-                },
-              ),
-              const SizedBox(height: 32),
-              SizedBox(
-                width: 300,
-                height: 56,
-                child: ElevatedButton(
-                  onPressed: () {
-                    if (_formKey.currentState!.validate()) {
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(
-                            builder: (context) => const ChatRoomPage()),
-                      );
-                    }
-                  },
-                  child: const Text(
-                    'Log In',
-                    style: TextStyle(fontSize: 16),
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class SignupPage extends StatelessWidget {
-  const SignupPage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-        appBar: AppBar(
-          backgroundColor: Colors.transparent,
-        ),
-        body: LoginForm());
   }
 }
