+++ lib/pages/signup_page.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+import 'package:balance_ai_agent/widgets/login_form.dart';
+
+class SignupPage extends StatelessWidget {
+  const SignupPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+        appBar: AppBar(
+          backgroundColor: Colors.transparent,
+        ),
+        body: const LoginForm());
+  }
+}
